requirements.txt
httpx>=0.27,<1
pytest>=8,<9
python-telegram-bot[job-queue]>=22.5,<23
